chore: Move contract interface creation to the useContractRegistry hook.
