+++ src/hooks/Guilds/contracts/useContractRegistry.ts
@@ -1,4 +1,5 @@
 import { useWeb3React } from '@web3-react/core';
+import { utils } from 'ethers';
 import { useMemo } from 'react';
 import useIPFSFile from '../ipfs/useIPFSFile';
 
@@ -27,19 +28,49 @@ export interface RegistryContract {
   tags: string[];
   networks: { [chainId: number]: string };
   functions: RegistryContractFunction[];
+  contractAddress: string;
+  contractInterface: utils.Interface;
 }
+
+type IPFSRegistryContract = Omit<
+  RegistryContract,
+  'contractAddress' | 'contractInterface'
+>;
 
 export const useContractRegistry = (chainId?: number) => {
   const { chainId: activeChainId } = useWeb3React();
 
-  const { data, error } = useIPFSFile<RegistryContract[]>(CONTRACT_REGISTRY);
+  const { data, error } =
+    useIPFSFile<IPFSRegistryContract[]>(CONTRACT_REGISTRY);
 
-  const registryContracts = useMemo(() => {
+  const registryContracts: RegistryContract[] = useMemo(() => {
     if (error || !data) return null;
 
-    return data.filter(
-      contract => !!contract.networks[chainId || activeChainId]
-    );
+    return data
+      .filter(contract => !!contract.networks[chainId || activeChainId])
+      .map(contract => {
+        // Construct the Ethers Contract interface from registry data
+        const contractInterface = new utils.Interface(
+          contract.functions.map(f => {
+            const name = f.functionName;
+            const params = f.params.reduce(
+              (acc, cur, i) =>
+                acc.concat(
+                  `${cur.type} ${cur.name}`,
+                  i === f.params.length - 1 ? '' : ', '
+                ),
+              ''
+            );
+            return `function ${name}(${params})`;
+          })
+        );
+
+        return {
+          ...contract,
+          contractAddress: contract.networks[chainId || activeChainId],
+          contractInterface,
+        };
+      });
   }, [chainId, activeChainId, data, error]);
 
   return {
